examples/quickstart-tensorflow-he/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tfheexample"
version = "1.0.0"
description = "Federated Learning with TensorFlow/Keras, Flower y TenSEAL (CKKS)"
license = "Apache-2.0"
dependencies = [
  "flwr[simulation]==1.19.0",
  "flwr-datasets[vision]==0.4.0",
  "tensorflow-cpu==2.15.0 ; platform_machine=='x86_64'",
  "tensorflow-macos>=2.9.1,!=2.11.1 ; platform_machine=='arm64'",
  "tenseal>=0.3.0",
  "matplotlib",
  "pandas",
]

[tool.hatch.build.targets.wheel]
packages = ["tfheexample"]

[tool.flwr.app]
publisher = "flwrlabs"

# App CKKS (TenSEAL): server y client de tfexample

[tool.flwr.app."quickstart-tensorflow-he/tfexample/server_app.py:app"]
client-app = "quickstart-tensorflow-he/tfexample/client_app.py:app"
num-supernodes = 10
run-config = { num-server-rounds = 20, local-epochs = 1, batch-size = 32, learning-rate = 0.005, fraction-fit = 0.5, fraction-evaluate = 1.0, verbose = false, seed = 44 }


[tool.flwr.federations]
default = "local-simulation"

[tool.flwr.federations.local-simulation]
options.num-supernodes = 10

[tool.flwr.federations.local-deployment]
address = "127.0.0.1:9093"
root-certificates = "./.cache/certificates/ca.crt"
